set up a class for the winery markers
