+++ TCWineTour/Marker.swift
@@ -7,3 +7,22 @@
 //
 
 import Foundation
+import GoogleMaps
+
+class Marker: GMSMarker {
+    let place: Winery
+    
+    init(place: Winery) {
+    self.place = place
+    super.init()
+            
+    position = place.position
+    icon = UIImage(named: "wineBottle")
+
+    }
+
+}
+//func setMarker(position: CLLocationCoordinate2D, title: String, map: GMSMapView) -> GMSMarker{
+//    
+//    return Winery.
+//}
