+++ TCWineTour/ViewController.swift
@@ -12,44 +12,51 @@ import GoogleMaps
 
 class ViewController: UIViewController {
     
-    var placesClient: GMSPlacesClient?
+//    var placesClient: GMSPlacesClient?
+//    let marker = GMSMarker()
+    
+    
     
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        let chateauChantal = "ChIJY5Cc60vTSk0R2VMNMeozAC4"
-        
-        let placesClient = GMSPlacesClient()
-        
-        placesClient.lookUpPlaceID(chateauChantal, callback: { (place: GMSPlace?, error: NSError?) -> Void in
-            if let error = error {
-                print("lookup place id query error: \(error.localizedDescription)")
-                return
-            }
-            
-            if let place = place {
-                print("Place name \(place.name)")
-                print("Place address \(place.formattedAddress)")
-                print("Place placeID \(place.placeID)")
-                print("Place attributions \(place.attributions)")
-            } else {
-                print("No place details for \(chateauChantal)")
-            }
-        })
+//        let chateauChantal = "ChIJY5Cc60vTSk0R2VMNMeozAC4"
+                
+        chateauChantal.getPalceInfo(chateauChantal.placeID)
 
         
-        var camera = GMSCameraPosition.cameraWithLatitude(44.865389,
+        let camera = GMSCameraPosition.cameraWithLatitude(44.865389,
             longitude: -85.520597, zoom: 11)
-        var mapView = GMSMapView.mapWithFrame(CGRectZero, camera: camera)
+        let mapView = GMSMapView.mapWithFrame(CGRectZero, camera: camera)
         mapView.myLocationEnabled = true
         self.view = mapView
         
-        var marker = GMSMarker()
-        marker.position = CLLocationCoordinate2DMake(44.865389, -85.520597)
-        marker.title = "Traverse City"
-        marker.snippet = "Michigan"
-        marker.map = mapView
+//        chateauChantal.setMarker(chateauChantal.position, title: "Chateau Chantal", map: mapView)
+        
+//        
+//        let chateauChantalMarker = WineryMarker(position: CLLocationCoordinate2DMake(44.865389, -85.520597), title: "Chateau Chantal")
+//        chateauChantalMarker.
+////
+//        let chataeuChanalmarker = GMSMarker()
+//
+//        
+//        chataeuChanalmarker.position = CLLocationCoordinate2DMake(44.865389, -85.520597)
+//        chataeuChanalmarker.title = "Chateau Chantal"
+//        chataeuChanalmarker.snippet = "Winery"
+//        chataeuChanalmarker.map = mapView
+
+
+        
+        
+        
+//        twoLads.position = CLLocationCoordinate2DMake(44.933800, -85.493137)
+//        twoLads.title = "2 Lads"
+//        twoLads.snippet = "Winery"
+//        twoLads.map = mapView
+////
+        
     }
+    
 }
